Move the agents related to RAG to a new module in the wiseagents.agents package
Fixes https://github.com/wise-agents/wise-agents/issues/85
