+++ tests/wiseagents/test_challenger.py
@@ -2,13 +2,14 @@ import os
 import threading
 
 import pytest
-from wiseagents import WiseAgent, WiseAgentMessage, WiseAgentRegistry
+from wiseagents import WiseAgentMessage, WiseAgentRegistry
+from wiseagents.agents import CoVeChallengerRAGWiseAgent
 from wiseagents.llm import OpenaiAPIWiseAgentLLM
 from wiseagents.transports import StompWiseAgentTransport
 from wiseagents.vectordb import Document, PGVectorLangChainWiseAgentVectorDB
 from wiseagents.wise_agent_impl import PassThroughClientAgent
 
-from wiseagents.wise_agent_impl import CoVeChallengerRAGWiseAgent
+
 
 cond = threading.Condition()
